+++ src/test/runTest.ts
@@ -1,6 +1,6 @@
-import * as path from 'path';
+import * as path from 'path'
 
-import { runTests } from 'vscode-test';
+import { runTests } from 'vscode-test'
 
 async function main() {
 	try {
@@ -15,8 +15,8 @@ async function main() {
 		// Download VS Code, unzip it and run the integration test
 		await runTests({ extensionDevelopmentPath, extensionTestsPath });
 	} catch (err) {
-		console.error('Failed to run tests');
-		process.exit(1);
+		console.error('Failed to run tests')
+		process.exit(1)
 	}
 }