+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
-## Unreleased changes
-* Fix test cases
+## 2.7.0 - October 14, 2020
+* (refs [#10](https://github.com/pascalre/vscode-yaml-sort/issues/10)) Use custom sort recursively on the whole file
+* New configuration `vscode-yaml-sort.useCustomSortRecursively` will use the custom sort keywords recursively on a file, when using custom sort
 
 ## 2.6.0 - September 28, 2020
 * (refs [#11](https://github.com/pascalre/vscode-yaml-sort/issues/11)) support line breaks